// src/Types.d.ts
export interface SchedulesProps {
  name: string;
  type: string;
  dayStart: number;
  dayEnd: number;
  time: {
    start: string;
    end: string;
  };
}